refactor(server): isolate append pandas implem
